+++ lib/drab/client.ex
@@ -29,7 +29,7 @@ defmodule Drab.Client do
       modules = [Drab.Core | commander.__drab__().modules] # Drab.Core is included by default
       templates = Enum.map(modules, fn x -> "#{Module.split(x) |> Enum.join(".") |> String.downcase()}.js" end)
 
-      access_session = commander.__drab__().access_session
+      access_session = commander.__drab__().inherit_session
       session = access_session 
         |> Enum.map(fn x -> {x, Plug.Conn.get_session(conn, x)} end) 
         |> Enum.into(%{})