+++ CHANGES.md
@@ -1,12 +1,13 @@
 ## 0.2.1
-Introduced Drab Session: the way to access (read-only) the Plug Session map in event handlers.
+Introduced Drab Session: the way to access (read-only) the Plug Session map in the Commander.
 
 Changes:
 * `use Drab.Endpoint` instead of `socket Drab.config.socket, Drab.Socket` in endpoint.ex
-* session access for specified session keys
+* Drab Session with value inheritance from the Plug Session
+* event handler must return socket, warning in the other case
 
 Fixes:
-* security (#2): checks token on each event call
+* security (#2): checks token in each event call to prevent tampering
 
 
 ## 0.2.0 (2017-01-22)
@@ -22,6 +23,7 @@ Changes:
 * renamed Drab.Templates to Drab.Template
 * JS: Drab is global, Drab.launch_event() is available
 
+
 ## 0.1.1
 
 Changes:
@@ -34,3 +36,7 @@ Changes:
 
 Fixes:
 * atom leaking issue (#1)
+
+
+## 0.1.0
+First public version. Very shy.